+++ NEWS.md
@@ -1,6 +1,6 @@
 # metatools 0.1.6
 * Breaking change: `combine_supp()` requires that the QNAM columns are not in the source dataset (#64)
-* Breaking change: `build_from_derived()` no longer requires all datasets with predecessors, but can run on a single dataset or multiple 
+* Breaking change: `build_from_derived()` no longer requires all datasets with predecessors, but can run on a single dataset or multiple. It will also now accept unnamed datasets. 
 * `combine_supp()` now allows multiple `QNAM` values to go to the same `IDVAR` (#63)
 * Allow supp data to be zero-row with `combine_supp()` (#45)
 * Enhance `check_ct_data()` so that `na_acceptable` can now accept a vector of variables. Also add new argument `omit_vars` to control if any variables should be skipped when doing controlled terminology checks (#57)